Re-organize homepage's components
I'll keep all components related to a particular page on such page's folder.
