+++ client/src/components/Home/Cover.js
